kubebuilder: 2.2.0
The current formula doesn't install the binary assets which kubebuilder
needs when running test suites.

This change makes the formula use the releases from github instead of
building kubebuilder from source. The releases contain all the needed
assets and duplicating the building of those to homebrew would be hard
to maintain.

This is a simpler and safer approach.